config reload

+++ restow.sh
@@ -11,7 +11,10 @@ for elem in $list; do
 	stow "$elem"
 done
 
-sudo rm /etc/default/grub
+if [ "$@" = "--remove"]; then
+	sudo rm /etc/default/grub
+fi
+
 sudo stow --target=/etc/default/ grub
 
 sudo stow --target=/etc/ keyd
